feat: Application api=workqueue
Of the attached design, this PR covers WorkerPool and Applications with `api=workqueue`. It does not cover the upcoming WorkStealer.

See test7 for an example. One creates one or more WorkerPools, and points them to an Application and Dataset.

Since this PR does not yet have the WorkStealer, there is some hard-coded hacking for now as far as work assignment (from Dataset to worker in a WorkerPool) is concerned.