feat: wire workqueue dashboard server-side to kubernetes live data
This updates the event structure so that the event sources pass back individual updates, rather than the entire model. This simplifies the server-side logic (it can simply pass back the data without needing to shadow any larger context). This also should make it easier to add time series views to the UI.