fix: kubectl application stream parsing fails due to whitespace in command
We were using a plain kubectl table, and splitting by whitespace. But one of the fields of an Application is its command, which surely will contain whitespace. This switches us to use a jsonpath, so we can control the field separator.

This also updates the Application UI component to display the supports gpu ui only if it is true