fix: LiveDashboard was not idempotent w.r.t. its EventSources
This resulted in the application going kinda dead shortly after the initial load:
- HMR not working
- React.lazy not working after a certain point
- static asset loading not working after a certain point

i think these were all connected. it was not that any one was broken, but that the entire stream of communication was broken.

This happens due to our use of StrictMode, which fabricates a duplicate constructor on the main component, in order to trigger idempotency bugs in a fail-fast kinda way. we were failing fast!

This PR re-enables some of the lazy loads we had previously disabled due to this bug.