feat: switch sidebar to use a Nav rather than the filter tree
This gives us a more conventional approach, more in line with how patternfly was designed.

With this update, only one of the three kinds of resources will be visible at a time.

With this update, filtering is temporarily disabled. We can restore it in the conventional patternfly spot atop the page body.