feat: use podman if it is installed
- improves support for podman
- install podman cli and create podman machine from dashboard init
- improved error handling for kubectl streaming
- switches hack/init.sh to use podman if docker is not already installed.
- improve podman build and kind image loading to avoid re-push
- add hack/nuke.sh which tears down the podman machine

Note: travis still runs with docker, as ibm travis still runs ubuntu 20.04, and there is no podman v4 support there. we rely on podman v4 for the `--rootful` option to `podman machine init`. this in turn is necessitated by our use of an internal NFS server. if we can figure out how to get nfs running rootless we could relax all of this, perhaps.