chore: refactor helm charts to use helm sub-charts more fully
Before, we had some of the sub-charts placed into platform/charts. This moves all of the inlined sub-charts up one directory, and properly expresses them as sub-chart dependencies in platform/Chart.yaml.

This also adds tags to the sub-charts to allow more flexible filtering/conditioning of installing.

This also updates the dashboard prebuilds to prebuild `defaults`.