feat: task simulator
This adds the ability to inject simulated tasks into a task queue.

It also cleans up the "refresh/update" behavior in the UI to initiate a rolling restart of the controllers.

This also fixes a bug in the workstealer -- its accounting of the unassigned was off by -1.

TODO: add a UI element to initiate a task simulator.