feat: distinguish taskqueue from dataset internally
We had been using "dataset" internally and "taskqueue" in the UI.

TODO: this does not yet address the conflation of terms in the kubernetes controllers and CRDs; it is focused only on the UI parts. this is why there are still some references to `dataset` in the dashboard -- the remaining CRD uses of that term.