fix: improve behavior when kind cluster does not exist (yet)
Previously, we were only fetching the kind cluster kubeconfig once. If we subsequently brought up a kind cluster, portions of the logic would not update to use the kubeconfig of the new cluster; similarly if the cluster were recycled during the session.

This also updates the ComputeTargets streamer to improve the handling situations where kind is not configured (e.g. no kind CLI, no kind cluster). The streamer will now report a placeholder so that the UI can present this fact to the user, rather than either erroring out, or hanging until the cluster exists...