fix: avoid relying on kind-exported kubeconfig
i don't think this even works... we don't want to use the kind-exported kubeconfig when accessing non-kind clusters. and kind already adds itself to the user's existing KUBECONFIG... and the kind-exported kubeconfig is specific to one kind cluster, but what if we have more than one?

so, instead, let's for now rely on the user's existing KUBECONFIG