fix: improvements for creating and using multiple ComputeTargets
- fixed workerpool controller, as it was always targetting the current cluster
- UI updates to fix a number of bugs in the creation and use of multiple compute targets
- re-initializing JaaS on an existing kind cluster would fail due to existence of cluster

Creating a WorkerPool against a non-manager ComputeTarget still fails. This is because the current logic creates the WorkerPool always in the management cluster, and then applies the AppWrappered resources in the target cluster. But the AppWrapper uid-references the WorkerPool, which doesn't work if the WorkerPool is in a different cluster (ownerReferences...)

Does this mean we should instead create the WorkerPool itself in the target cluster, and let the controllers do their thing always local to the cluster? If so, then one would need a special client to create WorkerPools, rather than just applying a kubeconfig.

Or do we update the logic to avoid ownerReferences when spanning clusters? If so, then we'll need to manage the cross-cluster deletion ourselves.

Or do we switch to kubestellar for this stuff?