feat: have Run controller manage queues, not UI client
we had been relying on the UI client (platform/dashboard) to manage the run <-> queue association; moreover, there was some leftover logic that was relating application to queue, rather than run to queue

this updates the controllers and dashboard to shift this burden into the controllers. now the UI can simply create a run, rather than creating a bunch of related entities and worrying about garbage collection. this is the purview of controllers!