test: test7c and related tests are erratic
the tests are waiting for all pods associated with the test (via the app.kubernetes.io/part-of label) to either be all-Ready or all not-Ready

test7c uses a helm-based workdispatcher. this runs a pod that eventually is not-Ready (by virtue of finishing normally, i.e. after it has dispatched all of its work). which leaves a combination of ready and not-ready. this is a recent regression, i think. coming from my big PR where i changed the part-of labels so as to include the workdispatcher in that label group

this also includes a tiny code cleanup to the workdispatcher.sh to clarify a variable name