fix: avoid nested polling in workstealer by use of inotifywait
This eliminates the polling in main.go. In its place, it uses inotifywait to tell when it should invoke the go code.

TODO: we are still using (from #1252) `rclone bisync`. But we know when the main.go code has updates to push, and so could in the future either use a pair of one-directional `rclone sync` (the remote->local would still need to poll; the local->remote could be triggered by changes from the main.go code); or we could update the main.go code to talk directly to the s3 api for changes.