feat: avoid rclone bisync in workstealer
This updates the main.go code to emit to stdout a stream of filepaths that need to be uploaded. Thereby we can avoid a polling bisync, and instead immediately upload the files as they change. We also avoid `rclone bisync` which is marked as experimental/do not use in production.