feat: shrinkwrap app from git/filesystem
Updates shrinkwap.sh to allow shrinkwrapping an application from a git URL or from a filepath in the local filesystem. Updates the tests to use this capability, i.e. shrinkwrap each test. This removes the need for separate test and watsonx helm charts. This PR dismantles some of those. We can refactor the rest later.