fix: avoid random assignment in workstealer
e.g. this can lead to a situation where rebalance (which just moves back to the unassigned box) then reassigns (randomly) back to one of the workers that already has too much work.
e.g. if a new worker comes online, we may not assign it any work

This also addresses live/deadworker rows not having a stable order in the output.