feat: update Application crd so that `code` is an array rather than string
with a proper `name` and `source` field for each code array item. previously, we were using mysterious logic to try to come up with a name for the code. now it is explicit, and now we allow for more than one source file to be injected.

BREAKING CHANGE: this is a breaking change to the Application API