feat: dispatcher should touch a donefile when it exits
covered by test7c and updates to tests/bin/helpers.sh to look for donefile

This feature isn't used for anything yet. Next, we can update the workstealer to observe the donefile and drop a kill file in the workers inbox. Then the watcher/main.go code can observe this and exit.