feat: stop using indexed kubernetes jobs for workerpool
We were already hacking this with `completions: 999999`, something we needed to allow for changing the `parallelism` factor on the fly.

Now, if we want to allow workers to gracefully exit... if we specify a non-null value for `completions`, the batchv1.Job controller will keep on creating pods, since we specified a completions... Completions is really for bounded work queues. We have an unbounded model.