feat: add --repo-secret command line flag
This removes the global charts/templates/github_ibm_com, which was an ibm-specific repo secret, and also a "global" resource, i.e. one not associated with either a workerpool, the workstealer, or a dispatcher.

In its place, this adds a `--repo-secret` command line flag, and updates the lowering logic to inject this as needed by any given component.