feat: local backend
- split deploy-tests.sh into compile.sh and up.sh
- `enqueue file --wait` results in workdispatcher entering restart loop
- update copy-out to use LUNCHPAIL_EXE
- add LUNCHPAIL_TARGET env var
- run instances should report 0 if no runs are found
- refactor helper.sh to allow 'wait for no instances' logic to be reused

TODO: the minio component distinguishes "internal" and "external" ports, but only uses the per-run port for the latter; this works for Kubernetes, but not for local. we'll need to use the per-run port for the "internal" port, too (i.e. run minio server asking it to use a port of our choosing rather than a fixed 9000). otherwise, it is not possible to have more than one run concurrently active locally!

Signed-off-by: Nick Mitchell <nickm@us.ibm.com>