feat: rename qin/copyin -> upload
This is to avoid confusion with the *other* way we have been using "copy in", to mean "copy in to the worker/pod". I think that other meaning is better, because it aligns with existing "copy-in/copy-out" terminology. What we really meant by qin is "upload from my laptop to the s3/queue storage".

Signed-off-by: Nick Mitchell <nickm@us.ibm.com>