test: reduce number of queue ls calls in test validators
For no good reason, we are issuing a large number of `queue ls` commands to validate that the stdout and stderr files exist. Since we have already fetched a list of *all* output files, we can more simply just consult that list. Also this uses the same (new) check on the .code file.

This also removes the unnecessary check for the dispatcher done file (which result in more `queue ls` calls). This is unnecessary, since the main point of checking for the dispatcher done file was to validate that the dispatcher eventually self-destructs, which we check separately.

Signed-off-by: Nick Mitchell <nickm@us.ibm.com>