feat: if user passes inputs on `up` command line, copy outputs back when done
This updates test7-up-dispatch for coverage, and adds support for a `post.sh` in tests that will be run to allow for custom post-completion validation.

This also fixes a latent bug in pkg/be/kubernetes/portforward, a race that resulted in portforward failure due to two goroutines trying to use the same port.

Signed-off-by: Nick Mitchell <nickm@us.ibm.com>