fix: up pipeline may hang
This is due to minio using a bucket notification without a one-time list. we already solved this in the Listen() impl, so this PR updates things to use that and to ditch the old code entirely.

this pr also adds some more verbose logging and fixes some error handling in a few places (missing checks for err != nil)

Signed-off-by: Nick Mitchell <nickm@us.ibm.com>