fix: first pass of cleaning up s3 paths: workstealer
this removes the complex combination of env vars and implicit logic to encode s3 queue paths in favor of passing explicit paths on the command line. this lets us consolidate the logic in the front-end. the runtime only uses the explicit paths as given.

TODO: this only cleans up the workstealer

Signed-off-by: Nick Mitchell <nickm@us.ibm.com>