feat: change t-shirt sizing and minSize to minMemory
This removes the notion of t-shirt size, which tied num workers to memory constraints. This never made much sense. Instead, let's focus on the memory constraints as they are correctness related. The others (workers, cores) are less likely to be application constraints.

TODO: somehow we need to figure out how to get cpu and VM core sizing constraints in there. maybe in the pool spec?

Signed-off-by: Nick Mitchell <nickm@us.ibm.com>